fix(ui): share popover changes
