+++ ui/src/features/Editor/components/TopBar/components/ActionBar/components/index.ts
@@ -1,2 +1,2 @@
 export { default as DeployDialog } from "./DeployDialog";
-export { default as ShareDialog } from "./ShareDialog";
+export { default as SharePopover } from "./SharePopover";
