+++ ui/src/lib/i18n/locales/en.json
@@ -125,13 +125,10 @@
   "Give your deployment a meaningful description...": "",
   "Are you sure you want to proceed?": "",
   "Deploy": "",
-  "Sharing": "",
-  "Not Sharing": "",
   "Share Project": "",
+  "Copy URL": "",
   "Share your project's workflow with anyone with the URL. This is limited access to reading the contents of the canvas.": "",
-  "Currently sharing": "",
-  "Currently not sharing": "",
-  "URL: ": "",
+  "Sharing": "",
   "Deploy project's workflow": "",
   "Additional actions": "",
   "Export Project": "",
