+++ ui/src/lib/i18n/locales/es.json
@@ -125,13 +125,10 @@
   "Give your deployment a meaningful description...": "Dale a tu despliegue una descripción significativa...",
   "Are you sure you want to proceed?": "¿Estás seguro de que deseas continuar?",
   "Deploy": "Desplegar",
-  "Sharing": "Compartiendo",
-  "Not Sharing": "No compartiendo",
   "Share Project": "Compartir proyecto",
+  "Copy URL": "",
   "Share your project's workflow with anyone with the URL. This is limited access to reading the contents of the canvas.": "Comparte el flujo de trabajo de tu proyecto con cualquier persona con la URL. Este es un acceso limitado para leer el contenido del lienzo.",
-  "Currently sharing": "Actualmente compartiendo",
-  "Currently not sharing": "Actualmente no compartiendo",
-  "URL: ": "URL: ",
+  "Sharing": "Compartiendo",
   "Deploy project's workflow": "Desplegar el flujo de trabajo del proyecto",
   "Additional actions": "Acciones adicionales",
   "Export Project": "Exportar proyecto",
