+++ ui/src/lib/i18n/locales/fr.json
@@ -125,13 +125,10 @@
   "Give your deployment a meaningful description...": "Donnez une description significative à votre déploiement...",
   "Are you sure you want to proceed?": "Êtes-vous sûr de vouloir continuer?",
   "Deploy": "Déployer",
-  "Sharing": "Partage",
-  "Not Sharing": "Ne partage pas",
   "Share Project": "Partager le projet",
+  "Copy URL": "",
   "Share your project's workflow with anyone with the URL. This is limited access to reading the contents of the canvas.": "Partagez le flux de travail de votre projet avec n'importe qui ayant l'URL. Cela donne un accès limité à la lecture du contenu du canevas.",
-  "Currently sharing": "Actuellement partagé",
-  "Currently not sharing": "Actuellement non partagé",
-  "URL: ": "URL: ",
+  "Sharing": "Partage",
   "Deploy project's workflow": "",
   "Additional actions": "Actions supplémentaires",
   "Export Project": "Exporter le projet",
