+++ ui/src/lib/i18n/locales/ja.json
@@ -125,13 +125,10 @@
   "Give your deployment a meaningful description...": "デプロイに意味のある説明を付けてください...",
   "Are you sure you want to proceed?": "続行しますか？",
   "Deploy": "デプロイ",
-  "Sharing": "共有",
-  "Not Sharing": "共有していません",
   "Share Project": "プロジェクトを共有",
+  "Copy URL": "URLをコピー",
   "Share your project's workflow with anyone with the URL. This is limited access to reading the contents of the canvas.": "プロジェクトのワークフローをURLを持つ誰でも共有できます。これはキャンバスの内容を読むための制限されたアクセスです。",
-  "Currently sharing": "現在共有中",
-  "Currently not sharing": "現在共有していません",
-  "URL: ": "URL: ",
+  "Sharing": "共有",
   "Deploy project's workflow": "プロジェクトのワークフローをデプロイ",
   "Additional actions": "追加アクション",
   "Export Project": "プロジェクトをエクスポート",
