+++ ui/src/lib/i18n/locales/zh.json
@@ -125,13 +125,10 @@
   "Give your deployment a meaningful description...": "为您的部署提供有意义的描述...",
   "Are you sure you want to proceed?": "你确定要继续吗？",
   "Deploy": "部署",
-  "Sharing": "共享",
-  "Not Sharing": "未共享",
   "Share Project": "共享项目",
+  "Copy URL": "",
   "Share your project's workflow with anyone with the URL. This is limited access to reading the contents of the canvas.": "使用URL与任何人共享项目的工作流。这是有限的访问权限，只能读取画布的内容。",
-  "Currently sharing": "当前共享",
-  "Currently not sharing": "当前未共享",
-  "URL: ": "URL: ",
+  "Sharing": "共享",
   "Deploy project's workflow": "",
   "Additional actions": "附加操作",
   "Export Project": "导出项目",
